// src/openapi_p/models/SlimIndex.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IndexType } from './IndexType';
export type SlimIndex = {
    type: IndexType;
    createdAt: number;
    id: number;
    title: string;
    total: number;
    uid: number;
};