// src/openapi_p/models/SlimUser.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Avatar } from './Avatar';
export type SlimUser = {
    avatar: Avatar;
    group: number;
    id: number;
    joinedAt: number;
    nickname: string;
    sign: string;
    username: string;
};